refactor(model)!: Move PURL-related extension functions to a separate file
Prepare for adding more PURL functionality.
