+++ plugins/reporters/asciidoc/src/main/resources/templates/asciidoc/vulnerability_report.ftl
@@ -17,7 +17,7 @@
     License-Filename: LICENSE
 --]
 
-[#assign PurlUtils = statics['org.ossreviewtoolkit.model.utils.PurlUtilsKt']]
+[#assign PurlUtils = statics['org.ossreviewtoolkit.model.utils.PurlExtensionsKt']]
 
 :publisher: OSS Review Toolkit
 [#assign now = .now]
